using sshfs: just a command for export export deb/doc/iso -c

+++ src/commands/export/deb.ts
@@ -9,12 +9,9 @@ export default class ExportDeb extends Command {
 
   static flags = {
     help: Flags.help({ char: 'h' }),
+    all: Flags.boolean({ char: 'a', description: 'export all archs' }),
     clean: Flags.boolean({ char: 'c', description: 'remove old .deb before to copy' }),
-    amd64: Flags.boolean({ description: 'export amd64 arch' }),
-    i386: Flags.boolean({ description: 'export i386 arch' }),
-    armel: Flags.boolean({ description: 'export armel arch' }),
-    arm64: Flags.boolean({ description: 'export arm64 arch' }),
-    all: Flags.boolean({ char: 'a', description: 'export all archs' })
+    verbose: Flags.boolean({ char: 'v', description: 'verbose' })
   }
 
   async run(): Promise<void> {
@@ -24,6 +21,8 @@ export default class ExportDeb extends Command {
     const Tu = new Tools()
     Utils.warning(ExportDeb.description)
     await Tu.loadSettings()
+
+    const echo = Utils.setEcho(flags.verbose)
 
     let script = ''
     let arch = Utils.eggsArch()
@@ -39,12 +38,14 @@ export default class ExportDeb extends Command {
     let cmd = `rm -f ${rmount}\n`
     cmd += `mkdir ${rmount}\n`
     cmd += `sshfs ${Tu.config.remoteUser}@${Tu.config.remoteHost}:${Tu.config.remotePathDeb} ${rmount}\n`
-    cmd += `rm -f ${rmount}/${Tu.config.filterDeb}${arch}\n`
+    if (flags.clean){
+      cmd += `rm -f ${rmount}/${Tu.config.filterDeb}${arch}\n`
+    }
     cmd += `cp ${Tu.config.localPathDeb}${Tu.config.filterDeb}${arch}  ${rmount}\n`
     cmd += `sync\n`
     cmd += `umount ${rmount}\n`
     cmd += `rm -f ${rmount}\m`
-    console.log(cmd)
-    await exec(cmd, { echo: false, capture: true })
+
+    await exec(cmd, echo)
   }
 }
